+++ packages/eslint-plugin-pf-codemods/src/rules/helpers/getFromPackage.ts
@@ -5,11 +5,11 @@ import {
   Statement,
   Directive,
   ExportNamedDeclaration,
-  ImportDefaultSpecifier,
   ImportSpecifier,
   ExportSpecifier,
 } from "estree-jsx";
 import { pfPackageMatches } from "./pfPackageMatches";
+import { ImportDefaultSpecifierWithParent } from "./interfaces";
 
 type Declarations = ImportDeclaration | ExportNamedDeclaration;
 type Specifiers = ImportSpecifier | ExportSpecifier;
@@ -91,7 +91,7 @@ export function getDefaultImportsFromPackage(
   context: Rule.RuleContext,
   packageName: string,
   componentName: string = ""
-): ImportDefaultSpecifier[] {
+): ImportDefaultSpecifierWithParent[] {
   const astBody = context.getSourceCode().ast.body;
 
   const importDeclarations = astBody.filter(
@@ -110,7 +110,7 @@ export function getDefaultImportsFromPackage(
           imp.source.value?.toString().includes(componentName)) &&
         imp.specifiers[0]?.type === "ImportDefaultSpecifier"
     )
-    .map((imp) => imp.specifiers[0]) as ImportDefaultSpecifier[];
+    .map((imp) => imp.specifiers[0]) as ImportDefaultSpecifierWithParent[];
 }
 
 export function getAllImportsFromPackage(