+++ packages/eslint-plugin-pf-codemods/src/rules/v6/emptyStateHeaderMoveIntoEmptyState/emptyStateHeader-move-into-emptyState.ts
@@ -7,15 +7,14 @@ import {
   Node,
 } from "estree-jsx";
 import {
+  checkMatchingJSXOpeningElement,
   getAttribute,
   getAttributeText,
   getAttributeValueText,
-  getChildElementByName,
+  getChildJSXElementByName,
   getDefaultImportsFromPackage,
   getExpression,
   getFromPackage,
-  includesImport,
-  nodeIsComponentNamed,
   getChildrenAsAttributeValueText,
   getRemoveElementFixes,
 } from "../../helpers";
@@ -149,17 +148,23 @@ module.exports = {
     const pkg = "@patternfly/react-core";
     const { imports } = getFromPackage(context, pkg);
 
-    if (!includesImport(imports, "EmptyState")) {
+    const emptyStateImport = imports.find(
+      (specifier) => specifier.imported.name === "EmptyState"
+    );
+
+    if (!emptyStateImport) {
       return {};
     }
 
     return {
       JSXElement(node: JSXElement) {
-        if (!nodeIsComponentNamed(node, "EmptyState")) {
+        if (
+          !checkMatchingJSXOpeningElement(node.openingElement, emptyStateImport)
+        ) {
           return;
         }
 
-        const header = getChildElementByName(node, "EmptyStateHeader");
+        const header = getChildJSXElementByName(node, "EmptyStateHeader");
         const emptyStateTitleTextAttribute = getAttribute(node, "titleText");
 
         if (!header && emptyStateTitleTextAttribute) {
@@ -168,7 +173,7 @@ module.exports = {
           return;
         }
 
-        const titleChild = getChildElementByName(node, "Title");
+        const titleChild = getChildJSXElementByName(node, "Title");
 
         if (
           (!header || header.type !== "JSXElement") &&
@@ -194,7 +199,7 @@ module.exports = {
           removeElements.push(titleChild);
         }
 
-        const emptyStateIconChild = getChildElementByName(
+        const emptyStateIconChild = getChildJSXElementByName(
           node,
           "EmptyStateIcon"
         );