+++ packages/pf-codemods/CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.5](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.1.4...@patternfly/pf-codemods@2.1.5) (2024-08-26)
+
+**Note:** Version bump only for package @patternfly/pf-codemods
+
+
+
+
+
+## [2.1.4](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.1.3...@patternfly/pf-codemods@2.1.4) (2024-08-26)
+
+**Note:** Version bump only for package @patternfly/pf-codemods
+
+
+
+
+
+## [2.1.3](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.1.2...@patternfly/pf-codemods@2.1.3) (2024-08-22)
+
+**Note:** Version bump only for package @patternfly/pf-codemods
+
+
+
+
+
+## [2.1.2](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.1.1...@patternfly/pf-codemods@2.1.2) (2024-08-21)
+
+**Note:** Version bump only for package @patternfly/pf-codemods
+
+
+
+
+
+## [2.1.1](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.1.0...@patternfly/pf-codemods@2.1.1) (2024-08-21)
+
+**Note:** Version bump only for package @patternfly/pf-codemods
+
+
+
+
+
+# [2.1.0](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.0.58...@patternfly/pf-codemods@2.1.0) (2024-08-20)
+
+
+### Features
+
+* **README:** update to include all flags ([#742](https://github.com/patternfly/pf-codemods/issues/742)) ([c362262](https://github.com/patternfly/pf-codemods/commit/c36226241f87ee34605bfaab4dfc664e0a1b5352))
+
+
+
+
+
+## [2.0.58](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.0.57...@patternfly/pf-codemods@2.0.58) (2024-08-14)
+
+**Note:** Version bump only for package @patternfly/pf-codemods
+
+
+
+
+
+## [2.0.57](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.0.56...@patternfly/pf-codemods@2.0.57) (2024-08-13)
+
+**Note:** Version bump only for package @patternfly/pf-codemods
+
+
+
+
+
+## [2.0.56](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.0.55...@patternfly/pf-codemods@2.0.56) (2024-08-12)
+
+**Note:** Version bump only for package @patternfly/pf-codemods
+
+
+
+
+
 ## [2.0.55](https://github.com/patternfly/pf-codemods/compare/@patternfly/pf-codemods@2.0.54...@patternfly/pf-codemods@2.0.55) (2024-08-06)
 
 **Note:** Version bump only for package @patternfly/pf-codemods